+++ pytoolsbrunojatoba/spam/main.py
@@ -1,0 +1,6 @@
+class EnviadorDeSpam:
+    def __init__(self, sessao, enviador):
+        self.sessao = sessao
+        self.envior = enviador
+    def enviar_emails(self, remetente, assunto, corpo):
+        pass
