+++ pytoolsbrunojatoba/tests/test_spam/conftest.py
@@ -1,0 +1,17 @@
+from pytoolsbrunojatoba.spam.db import Conexao
+import pytest
+
+
+@pytest.fixture(scope='session')
+def conexao():
+    conexao_obj = Conexao()
+    yield conexao_obj
+    conexao_obj.fechar()
+
+
+@pytest.fixture
+def sessao(conexao):
+    sessao_obj = conexao.gerar_sessao()
+    yield sessao_obj
+    sessao_obj.roll_back()
+    sessao_obj.fechar()
