+++ pytoolsbrunojatoba/tests/test_spam/test_envio_para_base_de_usuarios.py
@@ -1,0 +1,10 @@
+from pytoolsbrunojatoba.spam.enviador_de_email import Enviador
+from pytoolsbrunojatoba.spam.main import EnviadorDeSpam
+
+def test_envio_de_spam(sessao):
+    enviador_de_spam = EnviadorDeSpam(sessao, Enviador())
+    enviador_de_spam.enviar_emails(
+        'brjatoba@djangosolutions.com',
+        'DevPro Course',
+        'Check the amazings modules'
+    )
