+++ pytoolsbrunojatoba/tests/test_spam/test_usuarios.py
@@ -1,21 +1,4 @@
-from pytoolsbrunojatoba.spam.db import Conexao
 from pytoolsbrunojatoba.spam.modelos import Usuario
-import pytest
-
-
-@pytest.fixture(scope='session')
-def conexao():
-    conexao_obj = Conexao()
-    yield conexao_obj
-    conexao_obj.fechar()
-
-
-@pytest.fixture
-def sessao(conexao):
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
 
 
 def test_salvar_usuario(sessao):
